started main component with categories and posts sub components
