+++ src/components/main/categories/categories.css
@@ -1,0 +1,21 @@
+/* MAIN CATEGORIES */
+.main__categories {
+	border: solid 1px black;
+	height: 80vh;
+	width: 30%;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+
+@media only screen and (max-width: 600px) {
+	.main__categories {
+		width: 100%;
+	}
+}
+
+/* H2 TITLE */
+.h2__categories {
+	border: solid 1px greenyellow;
+	margin: 10px;
+}
