+++ src/components/main/main.css
@@ -1,6 +1,13 @@
+/* MAIN */
 .Main {
-    height: 80vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
+	height: 80vh;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
+
+@media only screen and (max-width: 600px) {
+	.Main {
+		flex-direction: column;
+	}
 }
